// backend/src/generated/models.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Upload: { input: any; output: any; }
};

export type AnalysisResponse = {
  __typename?: 'AnalysisResponse';
  atsScore: Scalars['Float']['output'];
  confidence?: Maybe<Scalars['Float']['output']>;
  keywords: Array<Scalars['String']['output']>;
  strengths: Array<Scalars['String']['output']>;
  suggestions: Array<Scalars['String']['output']>;
};
// ...
export type CreateResumeInput = {
  file: Scalars['Upload']['input'];
  filename: Scalars['String']['input'];
};